Separate adjacency list/matrix generation methods and the atomgraph classes into their own files
